ofVideoPlayer*: setVolume changed to take a float 0..1
. ofLogWarning() spammed in ofVideoPlayer and ofiPhoneVideoPlayer if passed-in
volume is > 1.0f
. also example in apps/devApps/videoPlayerSetVolumeExample

+++ libs/openFrameworks/video/ofVideoPlayer.h
@@ -49,7 +49,7 @@ class ofVideoPlayer : public ofBaseVideoPlayer,public ofBaseVideoDraws{
 		bool				getIsMovieDone();
 
 		void 				setPosition(float pct);
-		void 				setVolume(int volume);
+		void 				setVolume(float volume); // 0..1
 		void 				setLoopState(ofLoopType state);
 		int					getLoopState();
 		void   				setSpeed(float speed);
